Refactor Constancias logic
